+++ src/main/java/com/used/lux/dto/user/order/ProductOrderCancelDto.java
@@ -8,14 +8,14 @@ public record ProductOrderCancelDto(
         Long id,
         String userName,
         String productName,
-        int productPrice,
+        Long productPrice,
         String cancelTerm,
         LocalDateTime createdAt,
         String createdBy,
         LocalDateTime modifiedAt,
         String modifiedBy
 ) {
-    public static ProductOrderCancelDto of(Long id, String userName, String productName, int productPrice, String cancelTerm,
+    public static ProductOrderCancelDto of(Long id, String userName, String productName, Long productPrice, String cancelTerm,
                                  LocalDateTime createdAt, String createdBy, LocalDateTime modifiedAt, String modifiedBy) {
         return new ProductOrderCancelDto(id, userName, productName, productPrice, cancelTerm, createdAt,
                 createdBy, modifiedAt, modifiedBy);
