+++ src/main/java/com/used/lux/request/order/OrderCreateRequest.java
@@ -5,9 +5,9 @@ public record OrderCreateRequest(
         String phoneNumber,
         String address,
         String email,
-        int payment,
+        Long payment,
         String requestTerm) {
-    public static OrderCreateRequest of(String name, String phoneNumber, String address, String email, int payment, String requestTerm) {
+    public static OrderCreateRequest of(String name, String phoneNumber, String address, String email, Long payment, String requestTerm) {
         return new OrderCreateRequest(name, phoneNumber, address, email, payment, requestTerm);
     }
 }
